Revert existing libqb split
