--- internal/c/libqb/printer.h
@@ -1,2 +1,0 @@
-// forward references
-void sub__printimage(int32 i);
